+++ lib/jekyll/rp_logs/rp_page.rb
@@ -1,3 +1,5 @@
+require "forwardable"
+
 module Jekyll
   module RpLogs
     class Page